Move calculate_coordinate to algorithms/
Fixes #1367
