+++ data_structures/rectangle.hpp
@@ -28,7 +28,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifndef RECTANGLE_HPP
 #define RECTANGLE_HPP
 
-#include "coordinate_calculation.hpp"
+#include "../algorithms/coordinate_calculation.hpp"
 
 #include <boost/assert.hpp>
