+++ unit_tests/algorithms/geometry_string.cpp
@@ -29,7 +29,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <boost/test/unit_test.hpp>
 
 #include "../../algorithms/polyline_compressor.hpp"
-#include "../../data_structures/coordinate_calculation.hpp"
+#include "../../algorithms/coordinate_calculation.hpp"
 
 #include <osrm/coordinate.hpp>
