+++ unit_tests/data_structures/static_rtree.cpp
@@ -25,7 +25,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 */
 
-#include "../../data_structures/coordinate_calculation.hpp"
+#include "../../algorithms/coordinate_calculation.hpp"
 #include "../../data_structures/static_rtree.hpp"
 #include "../../data_structures/query_node.hpp"
 #include "../../data_structures/edge_based_node.hpp"
